+++ src/lib.rs
@@ -30,7 +30,7 @@ pub use value::Value;
 pub use memory::redis_store::{RedisMemory, RedisMemoryError};
 
 #[cfg(feature = "sql-mem")]
-pub use memory::sql_store::{SqlMemory, SqlMemoryError};
+pub use memory::sqlite_store::{SqlMemory, SqlMemoryError};
 
 #[cfg(any(test, feature = "test-utils"))]
 pub use encoding::{
